+++ docs/examples/Experimental.tsx
@@ -1,5 +1,5 @@
 /** @jsx jsx */
-import { Component } from 'react';
+import { useState } from 'react';
 import { jsx } from '@emotion/react';
 import { CSSObject } from '@emotion/serialize';
 import moment, { Moment } from 'moment';
@@ -183,73 +183,47 @@ const Option = (props: OptionProps<DateOption, false>) => {
 
 interface DatePickerProps {
   readonly value: DateOption | null;
-  readonly onChange: (value: DateOption | null) => void;
+  readonly onChange: (newValue: DateOption | null) => void;
 }
 
-interface DatePickerState {
-  readonly options: readonly (DateOption | CalendarGroup)[];
-}
+const DatePicker = (props: DatePickerProps) => {
+  const [options, setOptions] = useState(defaultOptions);
 
-class DatePicker extends Component<DatePickerProps, DatePickerState> {
-  state: DatePickerState = {
-    options: defaultOptions,
-  };
-  handleInputChange = (value: string) => {
+  const handleInputChange = (value: string) => {
     if (!value) {
-      this.setState({ options: defaultOptions });
+      setOptions(defaultOptions);
       return;
     }
     const date = chrono.parseDate(suggest(value.toLowerCase()));
-    if (date) {
-      this.setState({
-        options: [createOptionForDate(date), createCalendarOptions(date)],
-      });
-    } else {
-      this.setState({
-        options: [],
-      });
+    if (!date) {
+      setOptions([]);
+      return;
     }
+    setOptions([createOptionForDate(date), createCalendarOptions(date)]);
   };
-  render() {
-    const { value } = this.props;
-    const { options } = this.state;
-    return (
-      <Select<DateOption, false>
-        {...this.props}
-        components={{ Group, Option }}
-        filterOption={null}
-        isMulti={false}
-        isOptionSelected={(o, v) => v.some((i) => i.date.isSame(o.date, 'day'))}
-        maxMenuHeight={380}
-        onChange={this.props.onChange}
-        onInputChange={this.handleInputChange}
-        options={options}
-        value={value}
-      />
-    );
-  }
-}
 
-interface State {
-  readonly value: DateOption | null;
-}
+  return (
+    <Select<DateOption, false>
+      {...props}
+      components={{ Group, Option }}
+      filterOption={null}
+      isMulti={false}
+      isOptionSelected={(o, v) => v.some((i) => i.date.isSame(o.date, 'day'))}
+      maxMenuHeight={380}
+      onChange={props.onChange}
+      onInputChange={handleInputChange}
+      options={options}
+      value={props.value}
+    />
+  );
+};
 
-export default class Experimental extends Component<{}, State> {
-  state: State = {
-    value: defaultOptions[0] as DateOption,
-  };
-  handleChange = (value: DateOption | null) => {
-    this.setState({ value });
-  };
-  render() {
-    const { value } = this.state;
-    const displayValue =
-      value && 'value' in value ? value.value.toString() : 'null';
-    return (
-      <div>
-        <pre>Value: {displayValue}</pre>
-        <DatePicker value={value} onChange={this.handleChange} />
-      </div>
-    );
-  }
-}
+export default () => {
+  const [value, setValue] = useState<DateOption | null>(
+    defaultOptions[0] as DateOption
+  );
+
+  return (
+    <DatePicker value={value} onChange={(newValue) => setValue(newValue)} />
+  );
+};