+++ packages/react-select/src/creatable/index.ts
@@ -1,0 +1,2 @@
+export * from '../Creatable';
+export { default } from '../Creatable';