+++ packages/react-select/src/internal/DummyInput.tsx
@@ -1,6 +1,7 @@
 /** @jsx jsx */
 import { Ref } from 'react';
 import { jsx } from '@emotion/react';
+import { removeProps } from '../utils';
 
 export default function DummyInput({
   innerRef,
@@ -8,10 +9,20 @@ export default function DummyInput({
 }: JSX.IntrinsicElements['input'] & {
   readonly innerRef: Ref<HTMLInputElement>;
 }) {
+  // Remove animation props not meant for HTML elements
+  const filteredProps = removeProps(
+    props,
+    'onExited',
+    'in',
+    'enter',
+    'exit',
+    'appear'
+  );
+
   return (
     <input
       ref={innerRef}
-      {...props}
+      {...filteredProps}
       css={{
         label: 'dummyInput',
         // get rid of any default styles