+++ packages/react-select/src/utils.ts
@@ -367,3 +367,17 @@ export function multiValueAsValue<Option, IsMulti extends boolean>(
 ): OnChangeValue<Option, IsMulti> {
   return multiValue as OnChangeValue<Option, IsMulti>;
 }
+
+export const removeProps = <Props extends object, K extends string[]>(
+  propsObj: Props,
+  ...properties: K
+): Omit<Props, K[number]> => {
+  let propsMap = Object.entries(propsObj).filter(
+    ([key]) => !properties.includes(key)
+  );
+
+  return propsMap.reduce((newProps: { [key: string]: any }, [key, val]) => {
+    newProps[key] = val;
+    return newProps;
+  }, {}) as Omit<Props, K[number]>;
+};